+++ markdown_to_mrkdwn/__init__.py
@@ -2,5 +2,5 @@
 
 from .converter import SlackMarkdownConverter
 
-__version__ = "0.3.3"
+__version__ = "0.3.4"
 __all__ = ["SlackMarkdownConverter"]